move filepicker to input components
